% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftrl.Dataset.R
\name{dim.ftrl.Dataset}
\alias{dim.ftrl.Dataset}
\title{Dimensions of ftrl.Dataset}
\usage{
\method{dim}{ftrl.Dataset}(x)
}
\arguments{
\item{x}{Object of class \code{ftrl.Dataset}}
}
\description{
Returns a vector of numbers of rows and of columns in an \code{ftrl.Dataset}.
}
\details{
Note: since \code{nrow} and \code{ncol} internally use \code{dim}, they can also
be directly used with an \code{ftrl.Dataset} object.
}
