% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTRLProx.importance.R
\name{FTRLProx.importance}
\alias{FTRLProx.importance}
\title{FTRL-Proximal Linear Model Feature Importance Function}
\usage{
FTRLProx.importance(model)
}
\arguments{
\item{model}{a FTRL-Proximal linear model object.}
}
\value{
A \code{data.table} of the features used in the model with their weight
}
\description{
FTRLProx.importance showes the most important features of FTRL-Proximal Linear Model.
}
\examples{
library(data.table)
library(FeatureHashing)
data(ipinyou)
m.train <- FTRLProx_Hashing(~ 0 + ., ipinyou.train[, -"IsClick", with = FALSE],
                            hash.size = 2^13, signed.hash = FALSE, verbose = TRUE)
ftrl_model <- FTRLProx_train(m.train, y = as.numeric(ipinyou.train$IsClick),
                             family = "binomial",
                             params = list(alpha = 0.01, beta = 0.1, l1 = 1.0, l2 = 1.0),
                             epoch = 10, verbose = TRUE)
FTRLProx.importance(ftrl_model)
}

