% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Perf_Print}
\alias{Perf_Print}
\title{Performance Printing Function
This is used for printing model performance of each round}
\usage{
Perf_Print(Round, Name, Value)
}
\description{
Performance Printing Function
This is used for printing model performance of each round
}
\keyword{internal}
