% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftrl.Dataset.R
\name{print.ftrl.Dataset}
\alias{print.ftrl.Dataset}
\title{Print ftrl.Dataset}
\usage{
\method{print}{ftrl.Dataset}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{a object of \code{ftrl.Dataset}.}

\item{verbose}{whether to print mapping}

\item{...}{not currently used}
}
\description{
Print information about ftrl.Dataset
Currently it displays dimensions and presence of feature mapping.
}
\examples{
library(data.table)
library(FeatureHashing)
data(ipinyou)
f <- ~ 0 + BidID + IP + City + AdExchange + BiddingPrice + PayingPrice
m.train <- FTRLProx_Model_Matrix(f, ipinyou.train[, -"IsClick", with = FALSE],
                                 label = as.numeric(ipinyou.train$IsClick))

m.train
print(m.train, verbose = TRUE)

}
