% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFTRLProximal.R
\docType{package}
\name{rFTRLProximal}
\alias{rFTRLProximal}
\alias{rFTRLProximal-package}
\title{rFTRLProximal: FTRL-Proximal Algorithm}
\description{
An efficient C++ based implementation of "Follow The (Proximally) Regularized Leader"
online learning algorithm. This algorithm was proposed in McMahan et al. (2013)
<DOI:10.1145/2487575.2488200>.
}

