% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Subsetting dgCMatrix}
\usage{
slice(x, i)
}
\arguments{
\item{x}{a transposed \code{dgCMatrix}.}

\item{i}{logical expression indicating elements or rows to keep.}
}
\value{
A transposed \code{dgCMatrix} containing the subset of rows that are selected.
}
\description{
Returns subsets of a \code{dgCMatrix}.
}
