% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTRLProx_as_Matrix.R
\name{FTRLProx_as_Matrix}
\alias{FTRLProx_as_Matrix}
\title{Transform Data Frame to Sparse Design Matrix for FTRL-Proximal Model}
\usage{
FTRLProx_as_Matrix(data)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table}. The original data.}
}
\value{
an object of class "dgCMatrix"
}
\description{
Transform a data frame to sparse "design" matrix for FTRL-Proximal Algorithm.
Please always use this function to generate sparse matrix for training and prediction.
}
\examples{
library(data.table)
library(FeatureHashing)
data(ipinyou)
m.train <- FTRLProx_as_Matrix(ipinyou.train[, c("BiddingPrice", "PayingPrice"), with = FALSE])
hash.mapping(m.train)
}

