% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTRLProx_Matrix_Match.R
\name{FTRLProx_Matrix_Match}
\alias{FTRLProx_Matrix_Match}
\title{Transform Sparse Design Matrix to Assigned Schema}
\usage{
FTRLProx_Matrix_Match(x, default_name)
}
\arguments{
\item{x}{a \code{data.frame} or \code{data.table}. The original data.}

\item{default_name}{the feature name of design matrix}
}
\value{
an object of class "dgCMatrix"
}
\description{
Transform a sparse "design" matrix to assigned schema for FTRL-Proximal Algorithm.
}
\examples{
library(data.table)
library(FeatureHashing)
data(ipinyou)
f <- ~ 0 + BidID + IP + City + AdExchange + BiddingPrice + PayingPrice
m.train <- FTRLProx_Model_Matrix(f, ipinyou.train[, -"IsClick", with = FALSE])
m.train <- FTRLProx_Matrix_Match(m.train,
                                 c("AdExchange2", "AdExchange3", "BiddingPrice", "PayingPrice"))
hash.mapping(m.train)
}

